feat: use extendted language data from theme context
