+++ client-app/core/types/index.ts
@@ -1,3 +1,6 @@
+export * from "./theme-context";
+export * from "./language";
+
 import { CartAddressType, MemberAddressType, OrderAddressType } from "@core/api/graphql/types";
 import { LocationQueryValue } from "vue-router";
 
@@ -139,18 +142,4 @@ export interface IThemeConfigPreset {
   color_footer_bottom_text?: string;
   color_footer_bottom_link?: string;
   color_footer_bottom_link_hover?: string;
-}
-
-export interface IThemeContext {
-  baseUrl?: string;
-  storeId?: string;
-  storeName?: string;
-  language?: string;
-  availLanguages?: string[];
-  catalogId?: string;
-  currency?: string;
-  availCurrencies?: string[];
-  userId?: string;
-  userName?: string;
-  settings?: { key: string; value: unknown }[];
 }
