+++ client-app/core/types/language.ts
@@ -1,0 +1,8 @@
+export interface Language {
+  cultureName: string;
+  nativeName: string;
+  threeLetterLanguageName: string;
+  twoLetterLanguageName: string;
+  twoLetterRegionName: string;
+  threeLetterRegionName: string;
+}
