+++ client-app/main.ts
@@ -29,7 +29,7 @@ import "@/assets/styles/main.scss";
     legacy: false,
     globalInjection: true,
     fallbackLocale: "en-US",
-    locale: themeContext.language,
+    locale: themeContext.language?.cultureName,
   });
 
   // Create and mount application
